more cleanup, separated out a lot of commonly used structs for better decoupling
