+++ g_src/text/textlines.cpp
@@ -15,16 +15,10 @@
 #include "../_external/zlib/contrib/minizip/unzip.h"
 
 #include "../util/svector.hpp"
-using std::string;
-
 #include "../util/endian.hpp"
-
 #include "../files/files.hpp"
-
 #include "../render/enabler.hpp"
-
 #include "../text/textlines.hpp"
-
 #include "../util/basics.hpp"
 
 extern char filecomp_buffer[20000];
@@ -32,31 +26,32 @@ extern char filecomp_buffer2[80000];
 extern char filecomp_buffer_aux[20000];
 extern char filecomp_buffer2_aux[80000];
 
+void textlinesst::load_raw_to_lines(const char *filename){
 
-void textlinesst::load_raw_to_lines(const char *filename)
-{
   clean_lines();
 
-    //LOAD THE FILE
-  std::ifstream fseed(filename);
-  if(fseed.is_open())
-  {
-    string str;
+  std::ifstream fseed(filename);  // Load the File
+  if(fseed.is_open()){
 
-    while(std::getline(fseed,str))
-    {
-      long end=(long)str.length();
+    std::string str;
 
-      while(end>0)
-      {
+    while(std::getline(fseed,str)){
+
+      long end = (long)str.length();
+
+      while(end>0){
+
         if(isspace(str[end-1]))end--;
         else break;
+      
       }
 
       str.resize(end);
-
       if(str.length()>0)text.add_string(str);
+
     }
   }
+
   fseed.close();
+
 }
