+++ g_src/text/textlines.hpp
@@ -1,22 +1,20 @@
-#ifndef TEXTLINES_H
-#define TEXTLINES_H
+#ifndef DF_GSRC_TEXT_TEXTLINES_HPP
+#define DF_GSRC_TEXT_TEXTLINES_HPP
 
-struct textlinesst
-{
+struct textlinesst {
+
   stringvectst text;
 
+  void load_raw_to_lines(const char* filename);
 
-
-  void load_raw_to_lines(const char *filename);
-
-  void clean_lines()
-  {
+  void clean_lines(){
     text.clean();
   }
-  void copy_from(textlinesst &ot)
-  {
+
+  void copy_from(textlinesst& ot){
     text.copy_from(ot.text);
   }
+
 };
 
-#endif
+#endif // DF_GSRC_TEXT_TEXTLINES_HPP
